+++ lib/bandit/http1/error.ex
@@ -1,4 +1,0 @@
-defmodule Bandit.HTTP1.Error do
-  # Represents an error suitable for return as an HTTP status
-  defexception [:message, :status]
-end