+++ test/bandit/http1/request_test.exs
@@ -35,7 +35,7 @@ defmodule HTTP1RequestTest do
 
   describe "suppressing protocol error logging" do
     test "errors are not logged if so configured", context do
-      context = http_server(context, http_1_options: [log_protocol_errors: false])
+      context = http_server(context, http_options: [log_protocol_errors: false])
 
       output =
         capture_log(fn ->
@@ -684,12 +684,18 @@ defmodule HTTP1RequestTest do
          context do
       client = SimpleHTTP1Client.tcp_client(context)
 
-      Transport.send(
-        client,
-        "POST /short_body HTTP/1.1\r\nhost: localhost\r\ncontent-length: 5\r\n\r\nABC"
-      )
+      errors =
+        capture_log(fn ->
+          Transport.send(
+            client,
+            "POST /short_body HTTP/1.1\r\nhost: localhost\r\ncontent-length: 5\r\n\r\nABC"
+          )
 
-      assert {:ok, "200 OK", _, "OK"} = SimpleHTTP1Client.recv_reply(client)
+          assert {:ok, "200 OK", _, "OK"} = SimpleHTTP1Client.recv_reply(client)
+          Process.sleep(1100)
+        end)
+
+      assert errors =~ "(Bandit.HTTPError) Unable to read remaining data in request body"
     end
 
     def short_body(conn) do
@@ -789,13 +795,11 @@ defmodule HTTP1RequestTest do
             ]
           )
 
-          assert SimpleHTTP1Client.recv_reply(client)
-                 ~> {:ok, "400 Bad Request", list(), "HTTP method POST unsupported"}
-
+          assert SimpleHTTP1Client.recv_reply(client) ~> {:ok, "400 Bad Request", list(), ""}
           Process.sleep(100)
         end)
 
-      assert errors =~ "HTTP method POST unsupported"
+      assert errors =~ "(Bandit.HTTPError) HTTP method POST unsupported"
     end
 
     test "returns a 400 and errors loudly in cases where an upgrade is indicated but upgrade header is incorrect",
@@ -817,14 +821,12 @@ defmodule HTTP1RequestTest do
             ]
           )
 
-          assert SimpleHTTP1Client.recv_reply(client)
-                 ~> {:ok, "400 Bad Request", list(),
-                  "'upgrade' header must contain 'websocket', got [\"NOPE\"]"}
-
+          assert SimpleHTTP1Client.recv_reply(client) ~> {:ok, "400 Bad Request", list(), ""}
           Process.sleep(100)
         end)
 
-      assert errors =~ "'upgrade' header must contain 'websocket', got [\\\"NOPE\\\"]"
+      assert errors =~
+               "(Bandit.HTTPError) 'upgrade' header must contain 'websocket', got [\"NOPE\"]"
     end
 
     test "returns a 400 and errors loudly in cases where an upgrade is indicated but connection header is incorrect",
@@ -846,14 +848,12 @@ defmodule HTTP1RequestTest do
             ]
           )
 
-          assert SimpleHTTP1Client.recv_reply(client)
-                 ~> {:ok, "400 Bad Request", list(),
-                  "'connection' header must contain 'upgrade', got [\"NOPE\"]"}
-
+          assert SimpleHTTP1Client.recv_reply(client) ~> {:ok, "400 Bad Request", list(), ""}
           Process.sleep(100)
         end)
 
-      assert errors =~ "'connection' header must contain 'upgrade', got [\\\"NOPE\\\"]"
+      assert errors =~
+               "(Bandit.HTTPError) 'connection' header must contain 'upgrade', got [\"NOPE\"]"
     end
 
     test "returns a 400 and errors loudly in cases where an upgrade is indicated but key header is incorrect",
@@ -874,13 +874,11 @@ defmodule HTTP1RequestTest do
             ]
           )
 
-          assert SimpleHTTP1Client.recv_reply(client)
-                 ~> {:ok, "400 Bad Request", list(), "'sec-websocket-key' header is absent"}
-
+          assert SimpleHTTP1Client.recv_reply(client) ~> {:ok, "400 Bad Request", list(), ""}
           Process.sleep(100)
         end)
 
-      assert errors =~ "'sec-websocket-key' header is absent"
+      assert errors =~ "(Bandit.HTTPError) 'sec-websocket-key' header is absent"
     end
 
     test "returns a 400 and errors loudly in cases where an upgrade is indicated but version header is incorrect",
@@ -902,14 +900,12 @@ defmodule HTTP1RequestTest do
             ]
           )
 
-          assert SimpleHTTP1Client.recv_reply(client)
-                 ~> {:ok, "400 Bad Request", list(),
-                  "'sec-websocket-version' header must equal '13', got [\"99\"]"}
-
+          assert SimpleHTTP1Client.recv_reply(client) ~> {:ok, "400 Bad Request", list(), ""}
           Process.sleep(100)
         end)
 
-      assert errors =~ "'sec-websocket-version' header must equal '13', got [\\\"99\\\"]"
+      assert errors =~
+               "(Bandit.HTTPError) 'sec-websocket-version' header must equal '13', got [\"99\"]"
     end
 
     test "returns a 400 and errors loudly if websocket support is not enabled", context do
@@ -1575,7 +1571,8 @@ defmodule HTTP1RequestTest do
                {[:bandit, :request, :start], %{monotonic_time: integer()},
                 %{
                   connection_telemetry_span_context: reference(),
-                  telemetry_span_context: reference()
+                  telemetry_span_context: reference(),
+                  conn: struct_like(Plug.Conn, [])
                 }}
              ]
     end
@@ -1812,8 +1809,7 @@ defmodule HTTP1RequestTest do
                 %{
                   connection_telemetry_span_context: reference(),
                   telemetry_span_context: reference(),
-                  error: string(),
-                  status: 400
+                  error: string()
                 }}
              ]
     end
@@ -1833,8 +1829,7 @@ defmodule HTTP1RequestTest do
                 %{
                   connection_telemetry_span_context: reference(),
                   telemetry_span_context: reference(),
-                  error: "Header read timeout",
-                  status: 408
+                  error: "Header read timeout"
                 }}
              ]
     end
@@ -1854,6 +1849,7 @@ defmodule HTTP1RequestTest do
                 %{
                   connection_telemetry_span_context: reference(),
                   telemetry_span_context: reference(),
+                  conn: struct_like(Plug.Conn, []),
                   kind: :exit,
                   exception: %RuntimeError{message: "boom"},
                   stacktrace: list()